requirements.txt
tavily-python>=0.5.0
python-dotenv>=1.0.0
markdown>=3.5.0
xhtml2pdf>=0.2.11
